update linalg,py, test=develop
