+++ python/paddle/fluid/tests/unittests/test_cross_op.py
@@ -79,7 +79,7 @@ class TestCrossAPI(unittest.TestCase):
         with program_guard(Program(), Program()):
             x = fluid.layers.data(name='x', shape=[-1, 3])
             y = fluid.layers.data(name='y', shape=[-1, 3])
-            z = paddle.cross(x, y, dim=1)
+            z = fluid.layers.cross(x, y, dim=1)
             exe = fluid.Executor(fluid.CPUPlace())
             res, = exe.run(feed={'x': self.data_x,
                                  'y': self.data_y},
@@ -93,7 +93,7 @@ class TestCrossAPI(unittest.TestCase):
         with program_guard(Program(), Program()):
             x = fluid.layers.data(name='x', shape=[-1, 3])
             y = fluid.layers.data(name='y', shape=[-1, 3])
-            z = paddle.cross(x, y)
+            z = fluid.layers.cross(x, y)
             exe = fluid.Executor(fluid.CPUPlace())
             res, = exe.run(feed={'x': self.data_x,
                                  'y': self.data_y},
@@ -109,7 +109,7 @@ class TestCrossAPI(unittest.TestCase):
         with fluid.dygraph.guard():
             x = fluid.dygraph.to_variable(self.data_x)
             y = fluid.dygraph.to_variable(self.data_y)
-            z = paddle.cross(x, y)
+            z = fluid.layers.cross(x, y)
             np_z = z.numpy()
         expect_out = np.array([[-1.0, -1.0, -1.0], [2.0, 2.0, 2.0],
                                [-1.0, -1.0, -1.0]])
@@ -119,7 +119,7 @@ class TestCrossAPI(unittest.TestCase):
         with fluid.dygraph.guard():
             x = fluid.dygraph.to_variable(self.data_x)
             y = fluid.dygraph.to_variable(self.data_y)
-            z = paddle.cross(x, y, dim=1)
+            z = fluid.layers.cross(x, y, dim=1)
             np_z = z.numpy()
         expect_out = np.array([[0.0, 0.0, 0.0], [0.0, 0.0, 0.0],
                                [0.0, 0.0, 0.0]])
